+++ glade_forms/lib/src/validator/glade_validator.dart
@@ -5,11 +5,7 @@ import 'package:glade_forms/src/validator/part/satisfy_predicate_part.dart';
 import 'package:glade_forms/src/validator/validator_error/validator_error.dart';
 import 'package:glade_forms/src/validator/validator_instance.dart';
 
-typedef ValidateFunction<T> = GladeValidatorError<T>? Function(
-  T value, {
-  required InputDependencies dependencies,
-  Object? extra,
-});
+typedef ValidateFunction<T> = GladeValidatorError<T>? Function(T value);
 
 class GladeValidator<T> {
   List<InputValidatorPart<T>> parts = [];
@@ -25,23 +21,18 @@ class GladeValidator<T> {
   void clear() => parts = [];
 
   /// Checks value with custom validation function.
-  void custom(
-    ValidateFunction<T> onValidate, {
-    Object? key,
-    InputDependencies dependencies = const [],
-  }) =>
-      parts.add(CustomValidationPart(customValidator: onValidate, key: key, dependencies: dependencies));
+  void custom(ValidateFunction<T> onValidate, {Object? key}) =>
+      parts.add(CustomValidationPart(customValidator: onValidate, key: key));
 
   /// Checks value through custom validator [part].
   void customPart(InputValidatorPart<T> part) => parts.add(part);
 
   /// Checks that value is not null. Returns [ValueNullError] error.
   void notNull({OnValidateError<T>? devError, Object? key}) => custom(
-        (value, {required dependencies, extra}) => value == null
+        (value) => value == null
             ? ValueNullError<T>(
                 value: value,
                 devError: devError,
-                extra: extra,
                 key: key ?? GladeErrorKeys.valueIsNull,
               )
             : null,
@@ -51,16 +42,12 @@ class GladeValidator<T> {
   void satisfy(
     SatisfyPredicate<T> predicate, {
     OnValidateError<T>? devError,
-    InputDependencies dependencies = const [],
-    Object? extra,
     Object? key,
   }) =>
       parts.add(
         SatisfyPredicatePart(
           predicate: predicate,
-          devError: devError ?? (value, _) => 'Value ${value ?? 'NULL'} does not satisfy given predicate.',
-          extra: extra,
-          dependencies: dependencies,
+          devError: devError ?? (value) => 'Value ${value ?? 'NULL'} does not satisfy given predicate.',
           key: key,
         ),
       );