+++ glade_forms/lib/src/validator/part/input_validator_part.dart
@@ -1,17 +1,10 @@
-import 'package:glade_forms/src/core/core.dart';
 import 'package:glade_forms/src/validator/validator_error/glade_validator_error.dart';
 
 abstract class InputValidatorPart<T> {
   // ignore: no-object-declaration, key can be any object
   final Object? key;
 
-  final InputDependencies dependencies;
+  const InputValidatorPart({this.key});
 
-  const InputValidatorPart({required this.dependencies, this.key});
-
-  GladeValidatorError<T>? validate(
-    T value, {
-    required Object? extra,
-    InputDependencies dependencies = const [],
-  });
+  GladeValidatorError<T>? validate(T value);
 }