+++ glade_forms/lib/src/validator/validator_error/glade_validator_error.dart
@@ -3,44 +3,39 @@ import 'package:glade_forms/src/core/glade_input_error.dart';
 import 'package:glade_forms/src/validator/validator_error/value_null_error.dart';
 
 /// When validation failed but we already have propert [T] value.
-typedef OnValidateError<T> = String Function(T? value, Object? extra);
+typedef OnValidateError<T> = String Function(T? value);
 
 abstract class GladeValidatorError<T> extends GladeInputError<T> with EquatableMixin {
   /// Error message when translation is not used. Useful for development.
   // ignore: prefer-correct-callback-field-name, ok name
   final OnValidateError<T> devError;
 
-  // ignore: no-object-declaration, extra can be any object
-  final Object? extra;
-
   /// Value which triggered validation and returned this error.
   final T? value;
 
   @override
   // ignore: no-object-declaration, error can be anything (but typically it is string)
-  Object get error => devError(value, extra);
+  Object get error => devError(value);
 
-  String get devErrorMessage => devError(value, extra);
+  String get devErrorMessage => devError(value);
 
   @override
   List<Object?> get props =>
       // ignore: list-all-equatable-fields, wait for DCM update
-      [value, devError, extra, key, error, isConversionError, isNullError];
+      [value, devError, key, error, isConversionError, isNullError];
 
   GladeValidatorError({
     required this.value,
     OnValidateError<T>? devError,
-    this.extra,
     super.key,
   }) : devError = devError ??
-            ((v, _) =>
-                'Value "${v ?? 'NULL'}" does not satisfy validation. [This is default validation meessage. Consider to set `devErro` to cutomize validation errors]');
+            ((v) =>
+                'Value "${v ?? 'NULL'}" does not satisfy validation. [This is default validation meessage. Use `devError` to customize validation errors]');
 
-  factory GladeValidatorError.cantBeNull(T? value, {Object? extra, Object? key}) =>
-      ValueNullError<T>(value: value, key: key, extra: extra);
+  factory GladeValidatorError.cantBeNull(T? value, {Object? key}) => ValueNullError<T>(value: value, key: key);
 
   @override
   String toString() {
-    return devError(value, extra);
+    return devError(value);
   }
 }