+++ storybook/lib/debug_model.dart
@@ -38,16 +38,14 @@ class AgeRestrictedModel extends GladeModel {
       validator: (v) => (v
             ..notNull()
             ..satisfy(
-              (value, extra, dependencies) {
-                final vipContentInput = dependencies.byKey<bool>('vip-input');
-
-                if (!vipContentInput.value) {
+              (value) {
+                if (!vipInput.value) {
                   return true;
                 }
 
                 return value >= 18;
               },
-              devError: (_, __) => 'When VIP enabled you must be at least 18 years old.',
+              devError: (_) => 'When VIP enabled you must be at least 18 years old.',
               key: _ErrorKeys.ageRestriction,
             ))
           .build(),
@@ -62,8 +60,8 @@ class AgeRestrictedModel extends GladeModel {
 
         return devMessage;
       },
-      onChange: (info, dependencies) {
-        dependencies.byKey<bool>('vip-input').value = info.value >= 18;
+      onChange: (info) {
+        vipInput.value = info.value >= 18;
       },
     );
     vipInput = GladeInput.create(
@@ -71,12 +69,10 @@ class AgeRestrictedModel extends GladeModel {
       value: false,
       inputKey: 'vip-input',
       dependencies: () => [ageInput],
-      onChange: (info, dependencies) {
-        final age = dependencies.byKey<int>('age-input');
-
-        if (info.value && age.value < 18) {
+      onChange: (info) {
+        if (info.value && ageInput.value < 18) {
           groupEdit(() {
-            age.value = 18;
+            ageInput.value = 18;
           });
         }
       },