+++ glade_forms/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 2.0.0
 - **[Breaking]**: TextEditingController is no more created automatically. When TextEditingController is used, input's behavior is slightly changed. See README.md for full info.
 - **[Breaking]**: GladeInput's controller is now private. Use factory constructors to create input.
+- **[Breaking]**: `Extra` parameter removed
+- **[Breaking]**: `dependencies` are no loonger passed into `onChange` and in validator.
 
 ## 1.6.0
 - **Improvement**: GladeModelDebugInfo is more colorful and polished.